fix(python): reference isomorphism is broken within __init__
Within the `__init__` of a class, reading `self` back from the JS
process would result in a proxy object created for the parent type,
instead of returning `self`. This is because the `JSIIMeta` metaclass
only registered the instance in the rerefence map **after** the
constructor had returned.

This adds an additional registration point right after the `create` API
returned from the kernel, making the reference available as early as it
can possibly be known.

Fixes aws/aws-cdk#8262