fix(jsii): excessive "exclude" in "tsconfig.json"
In cases where both `jsii.tsc.outDir` and `jsii.tsc.rootDir` are set,
and `rootDir` happens to be a parent directory of `outDir`, the
generated `exclude` would mathc all files from `rootDir`, resulting in
an empty input set.

This makes sure to not emit an `exclude` declaration for `outDir` in
this case, so that input files do not get excluded.