chore(jsii): improved diagnostics for package.json
Produces source-linked diagnostic messages when inspecting the
`package.json` file for configuration values in the same way that
messages are generated for `.ts` files. This provides a nicer experience
as it allows to link multiple context entries (i.e: this `devDependency`
entry is redundant because of that `dependency` entry). It will also
integrate nicer with IDE problem markers.