fix(python): missing imports for certain keyword arguments
When a keyword argument was obtained from the super-interface of some
struct, imports required to bring the inherited properties' types in
scope were not emitted due to lack of traversing the struct hierarchies.

The fix is tested by introducing a method that accepts a kwargs-liftable
argument that extends another struct. Without this fix, tests in
`@jsii/python-runtime` would fail due to being unable to load any of the
modules. A test that expressedly leverages this new class and method was
however introduced for good measure.

Fixes #1809