fix(jsii): selective exports declarations are ignored
When specific symbols are exported from a module using the
*ExportDeclaration* syntax:

```ts
export { Foo } from './foo';
```

Those declarations would not be processed at all, because they lack the
`ts.ModifierFlags.Export` modifier (which is to be expected from a
declaration that **is** an export statement).

Added the necessary code (and a test) to ensure these statements are
correctly detected, and the exported symbols are correctly processed
into the output assembly.

Fixes #1818