fix(rosetta): variadic arguments may use kwargs syntax in Python
When processing function calls, the signature of the function was not
considered when it was available, and an object literal used as the
final argument of a variadic call would be rendered as keyword arguments
when this is actually not a correct way to render variadic values.

This change attempts to read the resolved signature of the function
beging called, in order to correctly render variadic values.

Fixes #1821