fix(rosetta): OOpsie -- couldn't find root file back on Windows
The root file was located by comparing file paths in a way that was not
adjusting path separators for Windows, and the result was inability to
match normalized paths with Unix-style ones.

This uses the `typescript.Program.getSourceFile` method instead, which
locates the source file and handles the path matching correctly for us.