chore: a bunch of Windows-specific issues
Fixing a bunch of locations where `/` was asserted to be present in file
paths, instead of using `path.join` to be Windows-compatible.

Additionally, replaced the `diff-test` implementation with a
`node`-based one, as `diff` is not readily available in Windows and the
substitute command behaves just too differently.

Finally, a couple of tests were shelling out to other node processes,
but doing so unsafely (forgetting to pass `shell: true`, for example),
so that the current `node` runtime is systematically being used.