feat(jsii): detect changing visibility when overriding
Since C# does not allow changing the visibility of a member when
overriding it, added a check to ensure nobody attempts to make a
protected member public when overriding it.

This is a somewhat common pitfall for users (as seen for example in
aws/aws-cdk#9616), and the check will allow faster detection and
correction (i.e: at compilation time instead of during `jsii-pacmak`).