feat(python): allow setting additional trove classifiers
Setting the `jsii.targets.python.classifiers` property in `package.json`
to an array of strings will trigger insertion of those values in the
`classifiers` of the generated python package.

Classifiers are de-duplicated and normalized in `setup.py`. Some
classifiers prefixes are considered "managed" (an appropriate value is
automatically determined for those based on the package's metadata), and
while configuring those will not cause `jsii-pacmak` to fail, a warning
will be emitted as manually specifying any of those is usually a
Bad Idea™:

- `Development Status ::`
- `License ::`
- `Operating System ::`
- `Typing ::`

Existing (i.e: valid) trove classifiers are mastered in the
[`trove-classifiers`](https://github.com/pypa/trove-classifiers) PyPA
package.

This contributes to aws/cdk-ops#393