chore(superchain): install awscli from pip instead of yum
The version of awscli in the yum repository appears to be somewhat outdated, and lacks
support for certain recent services and features. Moving to the recommended way to
install (using pip) will ensure the most recent release is installed every time the
image is re-built.

The recommended install/upgrade instructions are hosted at:
https://docs.aws.amazon.com/cli/latest/userguide/install-linux-al2017.html#install-amazon-linux-pip

Fixes #1905