chore: dial down integration test
Since this integration test is rather slow (generating all CDK in a
2-CPU instance), focus it on running `jsii-rosetta` and `jsii-pacmak`
only on the `monocdk-experiment` package (and it's future `aws-cdk-lib`
name, for good measure).

This hopefully is going to allow the tests to return *much* faster, at
least until we can safely offload this to a self-hosted github runner.