feat(golang): introduce Golang code generation (#1551)
* feat(golang): introduce Golang code generation

Adds a new `GoLang` code generation handler. The handler defines Go specific
generators for JSII submodules, interfaces, structs, and enums. The directory
structure of the generated code mirrors public submodules and namespaces defined
in the JSII module, ie, there is one file per namespace. The generated type
model is in progress and subject to change.

Adds documentation to `docs/configuration.md` about configuring the Go target.
Go code generation can be enabled by adding the `go` key to the `jsii` section
of your package.json.

```json
{
  "jsii": {
    "targets": {
      "go": {}
    }
  }
}
```

The Go target is not currently supported for production use. This is the first
iteration of code generation and does not define a target runtime library.

Relates to #83

Co-authored-by: Hsing-Hui Hsu <hsinghui@amazon.com>
Co-authored-by: Mitchell Valine <mitchellvaline@yahoo.com>