fix(pacmak): go bad local package imports
When generating import statements within go code, if the package was a
child of the package currently being generated, the import statement
would not include the module name.

When a package is part of a module though this caused invalid imports.
So for all package import statements moduleNames are now included when
provided in the jsii config.