refactor: make Go type members have consistent returnType getters (#2001)
* chore: use pointer receivers for getters and setters

* chore: rename references -> reference on methods/properties

* refactor: make GoType an abstract class

Makes `GoType` an abstract class extended by all the types previously
defined in the `ModuleType` union type. Also eliminates the need for the
GoEmitter interface.

* refactor: rename GoType#parent -> GoType#pkg

This change makes the lookup of package names more obvious and avoids
the stutter of having to call `parent.parent` from methods and
properties.

* refactor: enforce returnType getter on TypeFields

* refactor: rename TypeField -> GoTypeMember

* refactor: remove GoTypeMemberBase

By specifying returnType as a property on the GoTypeMember interface and
implementing the method as a getter on the classes that implement it, we
can get rid of using an inherited abstract class.