chore: make yarn-upgrade use "minor" target for prod deps
Production dependencies (i.e: dependencies that are not
`devDependencies`) can generally not be "safely" updated to the latest
(major) release without risking subtle breakage (due to behavior changes
that may not be caught by unit tests).

In order to have a safer general mode of operation, the
`yarn-upgrade.yml` workflow will now use the "minor" (instead of
"latest") target for any production dependency of any package in the
repository (a dependency is minor-pinned for the whole repository if any
of the packages in the repository declares a "production" dependency on
it).