chore(python): documentation improvements
Improved the documentability of generated Python code by reducing the
use of forward references through keeping track of types that were
already emitted; and by emitting a fully qualified python type name in
the remaining forward references instead of a (possibly ambiguous)
unqualified name.

This is expected to have a positive impact on the result of `autodocs`
when generating the `sphinx` docsite.