fix(go): null returns and pointer getters/setters
Fixes go method code generation to not emit a return statement when the
return type is `void`. Adds logic to fix go compile errors on
getters/setters that refer to recursive type properties. Meaning, if a
struct contains a property that the type of the struct itself, the type
of the property in Go is a pointer to the struct. Getters and setters
now reflect this.

Update test step to include compilation of all generated jsii-calc
modules since all compile errors are fixed.