chore: run integration test with node 10 (#2044)
It is currently broken because the latest version of `yargs` (16.0.3)
introduced an `exports` map in it's `package.json` that prevents access
to it's `package.json`. This is used by `decdk` when trying to load the
type system, and currently fails.

The breakage only occurs when using `node >= 12` (as the `exports`
map feature was introduced there), so this PR changes the node runtime
used to run integration tests from `node 12` to `node 10`.