chore: publish Docker image from GitHub actions instead of Travis
This gets rid of the last use-case we had for TravisCI, which can hence
be disabled. The general experience ought to be somewhat better as there
is no only a single master for all meaningful task executions, instead
of being scattered on two distinct platforms.

Since TravisCI was also the only place where "working directory clean"
was being asserted, this also adds that check to the "main" workflow.