fix(dotnet): abundant nullability warnings in generated code
The generated code had a number of quirks related to nullability
annotations (`?` and `!`) that resulted in sometimes abundant warnings
emitted during the compilation of the generated code. This makes the
necessary changes to correctly reflect that method arguments can be
`null` in the `DeptyBase` methods; and adds the necessary `?` and `!`
where they were previously missing.

This results in much fewer compilation warnings being emitted while
compiling `jsii-calc` and dependencies (the remaining warnings are
actually "legitimate" in the sense that they are the result of
us/users making questionable naming choices).