feat(go): embed jsii runtime application
Bundles the `@jsii/runtime` application within the go runtime library.
This ships the necessary files as byte slices in
`embeddedruntime.generated.go`. The `client` initializer will un-pack
this into a temporary directory (cleaned up by `client.Close()`) and run
that, unless the user specified the `JSII_RUNTIME` environment variable
(in which case, that command is used instead).

The `client` also assigns `JSII_AGENT` to a string representing the `go`
stack version used during build (`runtime.Version()`), the operating
system under which the application is running and the CPU architecture.
The string is formatted as `VERSION/OS/ARCH`.

A new `runtime.go` was introduced that manages a singleton `client`
instance and exports a `Close` function that can be invoked to
gracefully clean up the child process.