fix(jsii-diff): allow modification of external structs
We used to be checking the compatibility of structs in-line, as
a part of checking whether the call signature of a function is
compatible with its previous signature (i.e., checking that a
return type is only strengthened and an input type is only weakened).

This had a couple of downsides:

- Performance: we would check the same type multiple times if it
  occurred (deeply) in multiple call signatures.
- Locality: we would report the mismatch against the function signature
  instead of the changed type.
- Locality(2): if the change occurred in an `external` type but the
  consuming function was `stable`, the error ought to be suppressed but
  was reported instead.

The last reason recently cropped up in real life so that was a good
opportunity to rectify this long-standing architectural defect. I'm
afraid the result is a rather large refactoring; I apologize in advance,
but rest assured that everything will be better in the future, dear
reader.

New way of working:

* First identify all types in the assembly
* Crawl all types that have callables (methods, properties,
  initializers) and identify the types they use and whether they
  occur in input or output position.
* Validate all types, taking extra care to structurally compare
  structs based on whether they appear in input or output positions.

This only checks every struct once, report errors against the
struct itself and allows ignoring of structs that are marked `external`
or `experimental`.

While working on this change, it occurred to me that we were missing
a couple of validations that I added at the same time:

- In a `@subclassable` type, no method or property types can be changed
  *at all* (because C# won't allow it).
- A mutable property's type cannot be changed at all (it can
  both not be strengthened because it can be written, and not be weakened
  because it can be read).

Fixes #2064.