feat(pacmak/java): emitted proxy types extend base where possible
When generating interface proxies, and the proxied interface has exactly
one super-interface, it is possible for it's proxy to re-use the
implementation of the base's proxy by extending it, instead of repeating
the whole implementation.

Doing so reduces the amount of code that is generated, compiled and
released, making it faster to compile, and smaller to download.
Additionally this means proxies for interfaces that extend a single
foreign (i.e: from a dependency) interface are safer to use (the proxy
remains "fully valid" even if the resolved version of the dependency has
additional methods or properties).

The same approach can unfortunately not be used for abstract class
proxies, or for interfaces with more than one super-interface, since
Java does not allow multiple inheritance (making it impossible to
inherit from more than one base proxy class).

Fixes #2014