feat(pacmak): prerelease identifier support
When an NPM package built with `jsii` is tagged with a version that
includes a prerelease identifier (e.g: `1.2.3-alpha.1`), automatically
represent this as a pre-release version in generated packages, too. For
most currently supported languages, this simply means forwarding the
version untouched, however for Python, prerelease identifiers need to
be mapped into one of the supported tokens in PyPI (`.dev#`, `.a#`,
`.b#` or `.rc#`).

Fixes #2114