chore(tests): migrate jest config to jest.config.ts
This enables customizing behavior when `CI=true`, which is
conventionally the case on continuous integration systems, where
parallelization can be detrimental to performance due to use of budget
compute platforms.

Additionally, switched `jest` runner to `jest-circus` (which will
eventually become the default `jest` runner), as it promises to be
somewhat faster (at least, it did not degrade performance as far as I
could tell).

The `jest.config.ts` file is naturally excluded from all `npm` packages,
as it is a `.ts` file. It has also been excluded from TypeScript
compiler inputs (via `exclude` in `tsconfig.json` files) as it does not
need to be compiled with the projects (which would additionally cause
spurious `jest.config.d.ts` and `jest.config.js` files to be created).