fix(go): fix generation of readonly and static properties (#2133)
Previously, we were generating setters for readonly properties, as well as including static properties as part of a class interface and implementation. This fixes the property generation to omit the setter declarations/implementations for readonly properties. It also generates static properties as the package level.

Fixes #2093

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0