feat(pacmak/java): emit default interface implementations (#2076)
It should be possible to add properties or methods to interfaces that
are only meant to be returned by APIs (i.e: non-`@sublassable`) without
breaking extensions of those interfaces in other modules. Unfortunatley,
when such an interface is extended in another module, it's generated
proxy class would previously not have implementations for the new
member, creating the potential for a `NoSuchMethodException` at runtime.

This PR changes the code generation so that a new `Jsii$Default` nested
interface is generated for each interface, which contains default
implementations (that forward calls to the *jsii kernel*) for all of its
members.

> *Note:* members whose name collide with members declared on
> `java.lang.Object` cannot have `default` implementations, since these
> would always be overridden by `java.lang.Object` implementations.

The new `Jsii$Default` interface is implemented (and the default
implementations used) by the `Jsii$Proxy` class. When an interface
extends one or more bases, its `Jsii$Default` implementation extends
those of it's parent interfaces (and inherits their implementations).

In order to support backwards compatibility with libraries generated
with versions of `jsii-pacmak` that do not implement this feature, a
new `metadata.jsii.pacmak.hasDefaultInterfaces` entry is added to the
*jsii assembly* at compilation time. It's value is used to determine
whether `jsii-pacmak` can rely on the existence of the `Jsii$Default`
interface being present.

Finally, new elements were added to the Java runtime library for jsii:
- The `@Internal` annotation is used to designate API elements that are
  not meant for public usage
- The `IJsiiObject` interface is the base interface of all
  `Jsii$Default` interfaces, and is implemented by `JsiiObject`. Since
  interface members need be `public`, the methods declared by this
  interface have been maked `@Internal`, and were renamed from `jsiiFoo`
  to `$jsii$foo` (in order to completely remove the risk for name
  clashes with user-defined APIs)

Fixes #2014



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0