fix(jsii): TypeError: Cannot read property 'getJsDocTags' of undefined
When a trailing semicolon was left after a method body, the compiler
would choke with a `TypeError` due to attempting to read the JSDoc tags
of an `undefined` symbol.

This checks for this particular case and ignores the "class member" that
is represented by the dangling semicolon; and emits a diagnostic message
informing users the token could be removed.

Fixes #2098