chore: in-process generated code regression tests (#2167)
Instead of shelling out to `jsii-pacmak`, this runs the regression tests
in-process, enabling coverage collection from these. This provides much
better visibility in terms of the test coverage (which is >80% overall),
while previously it was ridiculously under-estimated.

In order to achieve this feat, I moved the bits of logic in
`bin/jsii-pacmak.ts` into the `lib/index.ts` file, so that this code can
conveniently be invoked from a test. This would constitute a first step
towards a programmatic API for `jsii-pacmak` in some future.

This happens to also make `jsii-pacmak` a tiny bit faster (through
some improvements I made in the `async` handling code while
I was there), although noting really material.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0