feat(dotnet,java): finalize kernel on VM shutdown
This change adds shutdown hook registrations in the .NET and Java
runtimes so that the `node` process hosting the JSII runtime is
gracefully terminated upon program end. This allows any "on exit" hooks
registered in the JavaScript process to properly run when the host
program ends. This is similar to what was done in #2100 for Python.

Additionally, removed the code that surrounded handling of the `STDERR`
stream from the Kernel process: there is no need to `PIPE` it, and
simply inheriting it (i.e: subprocess `STDERR` is the same stream as the
parent's) results in a more intuitive behavior, as `STDERR` output from
child processes surfaces "immediately" even if the parent process
somehow cannot forward it anymore (e.g: if it crashed or deadlocked).
This also means the `STDERR` forwarding thread is no longer needed.