feat(dotnet,java): kernel process inherits host's STDERR
Removed the code that surrounded handling of the `STDERR` stream from
the Kernel process: there is no need to PIPE it, and simply inheriting
it (i.e: subprocess STDERR is the same stream as the parent's) results
in a more intuitive behavior, as STDERR output from child processes
surfaces "immediately" even if the parent process somehow cannot
forward it anymore (e.g: if it crashed or deadlocked). This also means
the STDERR forwarding thread is no longer needed.