feat(rosetta): make assembly validation opt-in
Assembly validation (i.e: checking that the content of the `.jsii` file
correctly implements the `@jsii/spec` schema) can be slow and memory
intensive. It is however typically not of great use, since we should
normally be able to assume assemblies are valid.

Added a `--validate-assemblies` option (default `false`) to control
whether the JSON Schema validation is done or not.