feat(dotnet,java): finalize kernel process on VM shutdown (#2247)
This change adds shutdown hook registrations in the .NET and Java
runtimes so that the `node` process hosting the JSII runtime is
gracefully terminated upon program end. This allows any "on exit" hooks
registered in the JavaScript process to properly run when the host
program ends. This is similar to what was done in #2100 for Python.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0