feat(go): behavioral interface concrete types
Adds generation of empty structs for behavioral interfaces. These are to
be used by the runtime so that methods that return behavioral interfaces
have some known type that is able to be constructed and returned. These
may be able to be made private (non-exported) though they may be
referenced by other JSII modules that depend on them as necessitated by
the runtime.

Additionally, changes type signatures in places to always refer to the
interface type of JSII classes, interfaces, and structs. This is
required since JSII methods always return the interface implementation
of complex types in order to resolve runtime reference values. In order
for the returned values to be passed into other methods, the methods
themselves must accept the interface type and not the concrete type.