fix(go): Method call return reference an implMap
Fixes code generation for go method calls to correctly pass an implMap
where needed and a reference to the return value instead of a concrete
type by value. Adds test for method call with named return type in
TestReturnSpecialParam to cover this.

Fixes logic for `usesReflectionPackage` which only needs to know if the
type has methods or parameters since all getters and methods need to
reference `reflect.Type` regardless of the contents of the implMap.

Refactor runtime calls to not reimplement the emission logic for
implMaps.