chore: auto push go runtime
Adds action to auto push to the `jsii-runtime-go` repository whenever
master us updated. This will allow for testing against specific builds
of the go runtime between releases. This is especially useful during the
pre-stable phase as iteration is relatively rapid.

Fixes the name of the go-runtime module constant inside `jsii-pacmak` to
correspond to the published repo name.

Adds the `build-repo-dir` build script to @jsii/go-runtime that will
build the contents of the `jsii-runtime-go` repository for publishing.

Adds linting and formatting to the `.ts` files in
`@jsii/go-runtime/build-scripts` since they are growing and becoming
more complex.