fix(jsii): unknown error in Assembler._validateHeritageClauses
When the source code does not compile due to type checking errors, the
TypeScript program model may include "error type" placeholders, to allow
the compiler to provide feedback on subsequent code. The placeholder
types do not have a symbol (quite intentionally!), and will hence cause
a failure in the `Assembler.validateHeritageClauses` method.

This change makes the code more defensive so it does not crash on this
symptom, and does not report an additional error either (the TypeScript
compiler will have done this for us already).

The bug is more likely to hit users operating under `--watch`, as the
type checker is more likely to be able to insert placeholder types in
this case, as a previous version of the program might be used as a
crutch to understand the error's context... This makes this bug
difficult to reproduce in lab conditions, and is why this change does
not include a corresponding unit test.

Fixes #2349