fix(runtime): CPU thrashing when recovering from EAGAIN
The `node` runtime does not allow synchronous interactions with STDIN,
which is a problem because the jsii kernel **MUST ABSOLUTELY** have a
synchronous communication channel with the host process (otherwise, all
native overrides would result in an asynchronous **JavaScript** proxy,
which would break all **JavaScript** usage of those - since the original
code does not expect to have to deal with promises).

Unfortunately, we are not able to provision FIFOs (using `mkfifo`)
because those thiings do not exist on Windows. The closest thing to
FIFOs on Windows is IPC Channels, however `node` offers no synchronous
access to those either.

In order to solve this problem, we are leveraging `worker_threads` (a
feature that is available since around `node 9`, however gated by the
`--experimental-worker` feature flag in `node < 12` -- thankfully the
API has not really changed before it was stabilized, and we are in
control of how the `node` process is started).

The main thread (which the applications tarts on) create a `Worker`
thread and provides it with a shared memory buffer and the non-blocking
file descriptor (typically, of `process.stdin`, aka: `0`). The `Worker`
thread and main thread coordinate using the `Atomics` API, so that the
main thread is effectively blocked until the `Worker` thread has made
some data available to it (the `Worker` thread is then blocked until the
main thread has consumed all of that data).

Note that the `Atomics.wait` call suspends the whole agent that makes it
until the condition breaks or time-out is reached. This means absolutely
nothing else will happen on that thread during this sleep, including
processing other promises: from the moment code calls into the
`SyncReader.readSync` method, and until it has returned, no other
**JavaScript** code will be executed on the calling thread.