feat(runtime): use dedicated file descriptors for sync IPC
This replaces the `@jsii/runtime` package entry point so that it creates
a child `node` process with two new pipe file descriptors (at FD3 and
FD4). This means the `@jsii/runtime` now has two processes instead of
one:

1. A (new) "wrapper" process
2. The "core" process

The `@jsii/runtime` core process now uses `FD#3` for synchronous reading
of requests sent by the host to the `@jsii/kernel`, and `FD#4` to
synchronously write responses back to the host process. This alleviates
the need to resort to hot-polling around `EAGAIN` errors that are
symptomatic of attempting blocking operations on a non-blocking file
descriptor (aws/aws-cdk#12069). This works as the non-standard (`>2`)
file descriptors are not managed by `libuv`, and hence are not made
non-blocking at the start of `node`, allowing blocking operations to be
reliably used with those descriptors.

The `@jsii/runtime` core process was updated to use those two new file
descriptors instead of directly using `STDIN` and `STDOUT`, allowing
the **JavaScript** code to naturally use `process.stdout` and
`process.stderr` (as well as all other APIs that eventually use the same
file descriptors, such as `console.log` and `console.error`) without
causing corruption of the host-kernel communication channel as was
previously the case (#2406).

Data written directly to the `@jsii/runtime` core process' `FD#1`
(`STDOUT`) and `FD#2` (`STDERR`) is intercepted by the wrapper process,
and encoded in JSON objects before being sent to the host application
via the wrapper process' `STDERR` pipe. The host app is now reponsible
to maintain a thread that reads from the wrapper's `STDERR` pipe, decode
the JSON objects, then write the data to the appropriate host process'
STDIO stream to emulate direct operations as closely as possible.

Fixes aws/aws-cdk#12069
Fixes #2406